Fix CD sign
Physics texts use a different Fourier transform convention than
Engineering texts, so we need to flip our signs.

We also need to conjugate the frequency response of the compensating
filter, which we can do by conjugating the discrete time filter, as:
F{x*[n]} = X*(-ω)
and X(ω) is an even function.